refactor(architecture): re-own the skills webview content with its feature module (MOD-SKILL-MANAGEMENT wave)
The SHELL<->SKILL-MANAGEMENT 2-cycle ran through webviewSkillContent.ts: a
skills-dominated webview content module living in the shell while the skills
controller imported it back. Following the todos/prompts pattern, the file
moves to src/skills/ and the shared escapeAttribute helper sinks into the
kernel (src/webviewHtmlEscape.ts). The module gains a presentation role for
the moved file, entrypoints narrow to the consumed surface, and the kernel
dependency is declared. The remaining shell edge is type-only.

Measured: baseline 5 -> 4 fingerprints (the SHELL<->SKILL pair dissolves);
the SCC cluster ejects SKILL-MANAGEMENT (10 -> 9); waiver bijection holds.
Ledger flips MOD-SKILL-MANAGEMENT to strict. No behavior change.
