+++ .ci/architecture-debt-baseline.json
@@ -8,8 +8,7 @@
         "2:MOD-AI-SESSION-PROVIDER->MOD-AI-SESSION-RUNTIME",
         "2:MOD-AI-SESSION-PROVIDER->MOD-DASHBOARD-SHELL",
         "2:MOD-AI-SESSION-PROVIDER->MOD-WORKSPACE-IDENTITY",
-        "2:MOD-DASHBOARD-SHELL->MOD-SKILL-MANAGEMENT",
-        "scc:MOD-AI-SESSION-ATTENTION|MOD-AI-SESSION-CONTROL|MOD-AI-SESSION-CONVERSATION|MOD-AI-SESSION-PROVIDER|MOD-AI-SESSION-RUNTIME|MOD-DASHBOARD-SHELL|MOD-OPEN-WORKSPACE|MOD-PROJECT-CATALOG|MOD-SKILL-MANAGEMENT|MOD-WORKSPACE-IDENTITY"
+        "scc:MOD-AI-SESSION-ATTENTION|MOD-AI-SESSION-CONTROL|MOD-AI-SESSION-CONVERSATION|MOD-AI-SESSION-PROVIDER|MOD-AI-SESSION-RUNTIME|MOD-DASHBOARD-SHELL|MOD-OPEN-WORKSPACE|MOD-PROJECT-CATALOG|MOD-WORKSPACE-IDENTITY"
       ]
     }
   }
