+++ docs/testing/architecture-program.json
@@ -107,10 +107,13 @@
             "nextAction": "awaiting its Stage 5 wave"
         },
         "MOD-SKILL-MANAGEMENT": {
-            "state": "legacy",
-            "since": "initial coarse registry (Stage 1A/2)",
-            "evidence": [],
-            "nextAction": "awaiting its Stage 5 wave"
+            "state": "strict",
+            "since": "Stage 5 wave (2026-08-20): the shell 2-cycle broke by re-owning webviewSkillContent.ts with its feature module (the todos/prompts pattern) and sinking the shared escapeAttribute helper into the kernel; remaining shell edge is type-only; entrypoints narrowed to the consumed surface",
+            "evidence": [
+                "tests/contract/skills/webviewSkillContent.test.js",
+                "tests/contract/skills/skillScanCache.test.js"
+            ],
+            "nextAction": "none — strict; hold via the architecture policy checks"
         },
         "MOD-TODO": {
             "state": "strict",
